+++ examples/echo_reply_edit_method.v
@@ -10,8 +10,8 @@ struct App {
 
 // handle all messagess
 [message]
-fn (mut app App) all_messages(result vt.Update) ! {
-	message := result.message.reply(mut app, text: '${result.message.text}')!
+fn (mut app App) all_messages(result vt.Result) ! {
+	message := result.update.message.reply(mut app, text: '${result.update.message.text}')!
 	time.sleep(1000 * time.millisecond)
 	message.edit(mut app, text: 'new text ${rand.int()}')!
 }
@@ -20,5 +20,6 @@ fn main() {
 	mut app := App{
 		token: '5401623750:AAFWXZWx8V-SZIDQUI62AT7agCMs55aLIdU'
 	}
-	vt.start_polling(mut app) // starting the bot
+	polling_config := vt.PollingConfig[vt.Regular]{}
+	vt.start_polling(mut app,polling_config) // starting the bot
 }